crawling tests for common docs sites
